Space Invaders: Multiplayer support - spawn enemies in appropriate locations.
Four enemies are spawned per damage received. They prefer to spawn
in the lowest row available, unless it is too close to the bottom.
They also prefer to spawn centrally within the existing cohort of
aliens. If there is no space, they will add new rows above (to the
point of spawning above the top of the screen if required).